# GCD based inversion for 31 bit fields