Following #920 I was wondering if a similar approach could speed up inversions for our 31 bit fields. Turns out the answer is yes!

The following gcd based inversion algorithm gives over a `2x` speed up for inversions of MontyField31's and Mersenne31 (along with their extensions):

|Field | Old Inv Speed (ns) | New Inv Speed (ns) |
| --- | --- | --- |
| BabyBear | 346 | 131 |
| KoalaBear | 343 | 128 |
| Mersenne31 | 183 | 91|
| BabyBear Deg 4 Ext | 475 | 167 |
| BabyBear Deg 5 Ext | 673 | 379 |
| KoalaBear Deg 4 Ext | 477 | 163 |
| Mersenne31 Deg 4 Ext | 322 | 177 |
| Mersenne31 Deg 6 Ext | 432 | 291 |